"bottomColor" color sensor can be used
discovered that colorsensor only works if only one is attached. WIth two
colorsensors, they can either be both active or passive, and they seems
to share RGB value. So I removed beacon sensor physically and also in
the configuration. bottomColor starts to work. the code includes a
psuedo code for line following.